fix output of debug
the new package loader actually loads in packages that don't exist even if
they are result of an error, previously the loader didn't seem to have these
in the set of all packages returned though I couldn't figure out logically how
it happened from reading the code. in any case, slight change to printing to
make this more clear and update the test slightly.

tests still failing at refactorings that should error but do not