fix up tests to run w/ modules off
the GO111MODULES env var is going away in 1.17 so we don't have this luxury
really for very long, but it would be nice to get tests working prior to
migrating things so this is gonna be how it is.

TODO I don't think we can set the GO111MODULE flag where it is now, need to do
it only for tests so that users can actually use modules
TODO cgo is broken
TODO a few extract func tests seem broken, the nature of them seems that the
behavior of the type checking has changed or the extract func refactoring
itself was looking for some specific behaviors that have now broken, but in
any case, the failing tests themselves the refactorings produce edit sets that
would make the changes that should fail so it's not a loading issue or a
testing issue, there's a behavior thing to look into and debug.

^ only remaining issues afaik